Rename directory in deployEngine._on_deploy method and add test for rolling update.(#12)

+++ iconservice/deploy/icon_score_deploy_util.py
@@ -21,5 +21,6 @@ class DirectoryNameConverter:
 
     @classmethod
     def rename_directory(cls, path: str):
-        cls.counter += 1
-        os.rename(path, f"{path}{int(time.time()*10**6)}{cls.counter}_garbage_score")
+        if os.path.exists(path):
+            cls.counter += 1
+            os.rename(path, f"{path}{int(time.time()*10**6)}{cls.counter}_garbage_score")
